Rename folder holding modals to Modal
